Implement Realms for app
Made Quote and Watchlist like Objects in Realms.
Made WatchlistsDataSource to fetch and rule of Watchlists list in DB.
Made updating data in DB and updating views from DB.
EditEnabledDiffableDataSource in WatchlistViewController now has a Quote instead of String in generic.

+++ Quoteslist/Views/EditEnabledDiffableDataSource.swift
@@ -35,12 +35,14 @@ where MyType: Hashable {
         guard let id = itemIdentifier(for: indexPath) else {
             return
         }
-        var currentSnapshot = self.snapshot()
-        currentSnapshot.deleteItems([id])
+        if editingStyle == .delete {
+            var currentSnapshot = self.snapshot()
+            currentSnapshot.deleteItems([id])
 
-        self.apply(currentSnapshot)
+            self.apply(currentSnapshot)
 
-        self.deleteClosure?(id)
+            self.deleteClosure?(id)
+        }
     }
 
     override func tableView(
